refactor: 코드 및 패키지 정리
