+++ src/main/java/com/programmers/week/customer/application/CustomerService.java
@@ -1,11 +1,11 @@
 package com.programmers.week.customer.application;
 
-import com.programmers.week.Message;
-import com.programmers.week.customer.ui.CustomerCreateRequest;
 import com.programmers.week.customer.domain.Customer;
 import com.programmers.week.customer.infra.CustomerRepository;
-import com.programmers.week.customer.ui.CustomerResponse;
-import com.programmers.week.customer.ui.CustomerUpdateRequest;
+import com.programmers.week.customer.presentation.CustomerCreateRequest;
+import com.programmers.week.customer.presentation.CustomerResponse;
+import com.programmers.week.customer.presentation.CustomerUpdateRequest;
+import com.programmers.week.exception.Message;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -17,28 +17,28 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class CustomerService {
 
-  private final CustomerRepository customerRepository;
+    private final CustomerRepository customerRepository;
 
-  @Transactional
-  public Long create(CustomerCreateRequest request) {
-    Customer customer = new Customer(
-            request.firstName(),
-            request.lastName()
-    );
-    Customer saved = customerRepository.save(customer);
-    return saved.getId();
-  }
+    @Transactional
+    public Long create(CustomerCreateRequest request) {
+        Customer customer = new Customer(
+                request.firstName(),
+                request.lastName()
+        );
+        Customer saved = customerRepository.save(customer);
+        return saved.getId();
+    }
 
-  public CustomerResponse findById(Long id) {
-    return customerRepository.findById(id)
-            .map(CustomerResponse::from)
-            .orElseThrow(() -> new IllegalArgumentException(Message.CUSTOMER_IS_NO_EXIST));
-  }
+    public CustomerResponse findById(Long id) {
+        return customerRepository.findById(id)
+                .map(CustomerResponse::from)
+                .orElseThrow(() -> new IllegalArgumentException(Message.CUSTOMER_IS_NO_EXIST));
+    }
 
-  public Page<CustomerResponse> findAll(PageRequest pageRequest) {
-    Page<Customer> customerPage = customerRepository.findAll(pageRequest);
-    return customerPage.map(CustomerResponse::from);
-  }
+    public Page<CustomerResponse> findAll(PageRequest pageRequest) {
+        Page<Customer> customerPage = customerRepository.findAll(pageRequest);
+        return customerPage.map(CustomerResponse::from);
+    }
 
   @Transactional
   public Long update(CustomerUpdateRequest request) {
@@ -49,9 +49,9 @@ public class CustomerService {
     return customer.getId();
   }
 
-  @Transactional
-  public void deleteById(Long id) {
-    customerRepository.deleteById(id);
-  }
+    @Transactional
+    public void deleteById(Long id) {
+        customerRepository.deleteById(id);
+    }
 
 }
