--- src/main/java/com/programmers/week/customer/intfra/CustomerRepository.java
@@ -1,9 +1,0 @@
-package com.programmers.week.customer.intfra;
-
-import com.programmers.week.customer.domain.Customer;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface CustomerRepository extends JpaRepository<Customer, Long> {
-}
