+++ src/main/java/com/programmers/week/customer/presentation/CustomerCreateRequest.java
@@ -1,0 +1,7 @@
+package com.programmers.week.customer.presentation;
+
+public record CustomerCreateRequest(
+        String firstName,
+        String lastName
+) {
+}
