+++ src/main/java/com/programmers/week/customer/presentation/CustomerUpdateRequest.java
@@ -1,0 +1,7 @@
+package com.programmers.week.customer.presentation;
+
+public record CustomerUpdateRequest(
+        String firstName,
+        String lastName
+) {
+}
