+++ src/main/java/com/programmers/week/customer/presentation/PageRequestDto.java
@@ -1,0 +1,7 @@
+package com.programmers.week.customer.presentation;
+
+public record PageRequestDto(
+        int page,
+        int size
+) {
+}
