--- src/main/java/com/programmers/week/customer/ui/CustomerCreateRequest.java
@@ -1,4 +1,0 @@
-package com.programmers.week.customer.ui;
-
-public record CustomerCreateRequest(String firstName, String lastName) {
-}
