--- src/main/java/com/programmers/week/customer/ui/CustomerUpdateRequest.java
@@ -1,4 +1,0 @@
-package com.programmers.week.customer.ui;
-
-public record CustomerUpdateRequest(Long id, String firstName, String lastName) {
-}
