--- src/main/java/com/programmers/week/customer/ui/PageRequestDto.java
@@ -1,4 +1,0 @@
-package com.programmers.week.customer.ui;
-
-public record PageRequestDto(int page, int size) {
-}
