+++ src/main/java/com/programmers/week/exception/Message.java
@@ -1,4 +1,4 @@
-package com.programmers.week;
+package com.programmers.week.exception;
 
 public class Message {
 
@@ -9,9 +9,10 @@ public class Message {
   public static final String CHEF_IS_NULL = "요리사가 비어있습니다.";
   public static final String MEMO_IS_NULL = "메모가 없습니다.";
 
-  public static final String POWER_PRICE_IS_MINUS = "파워가 음수입니다. : ";
-  public static final String TOTAL_PRICE_IS_MINUS = "가격이 음수입니다. : ";
-  public static final String TOTAL_QUANTITY_IS_MINUS = "재고가 음수입니다. : ";
+  public static final String POWER_PRICE_IS_MINUS = "범위를 벗어나는 파워를 입력하였습니다. : ";
+  public static final String TOTAL_PRICE_IS_MINUS = "범위를 벗어나는 가격을 입력하였습니다. : ";
+  public static final String TOTAL_QUANTITY_IS_MINUS = "범위를 벗어나는 수량을 입력하였습니다. : ";
+  public static final String NAME_LENGTH_IS_WRONG = "범위를 벗어나는 길이의 이름을 입력하였습니다. : ";
 
   public static final String INCORRECT_ORDER_STATUS = "수정할 주문 상태가 올바르지 않습니다. ";
