+++ src/main/java/com/programmers/week/item/domain/Car.java
@@ -1,8 +1,7 @@
 package com.programmers.week.item.domain;
 
-import com.programmers.week.Message;
-import jakarta.persistence.DiscriminatorValue;
-import jakarta.persistence.Entity;
+import com.programmers.week.exception.Message;
+import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
