+++ src/main/java/com/programmers/week/item/domain/Food.java
@@ -1,9 +1,7 @@
 package com.programmers.week.item.domain;
 
-import com.programmers.week.Message;
-import jakarta.persistence.Column;
-import jakarta.persistence.DiscriminatorValue;
-import jakarta.persistence.Entity;
+import com.programmers.week.exception.Message;
+import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
